Almost compiles

+++ core/lib/storage/migrations/2021-09-03-163738_no-2fa-pubkey-hash/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE no_2fa_pub_key_hash
 (
-    account_id INTEGER PRIMARY KEY,
+    account_id BIGINT PRIMARY KEY,
     pub_key_hash TEXT NOT NULL
 );
